reorganize testHelpers
